fixes in for encrypted commit log: - correctly write out the  END_OF_SEGMENT marker. Was previously calling RAR.write(int), which only  write out a byte, not an int. Now calling RAR.writeInt(int)
- when reading an encrypted file, for each block we read the four byte header which indicates the length of the following (encrypted) block. If there's less than four bytes left
in the channel (meaning, you are almost at the end of the file), check the number of bytes read to see if it's less than four - and if so, the file is done.

- In ESW, add an extra call to sync() when close()'ing the file, just to make sure there's nothing left in the pre-encrypt buffer (that needs to get encrypted and flushed).

- fix for commit log archiving and encrypted commit logs. The problem was the encrypted commit log was not flushed before
the commit log archive command executes, and thus you miss some commits when replaying (uncovered by dtests). The solution
is to sync the segment in the archive executor. This should be safe as CLS.sync() checks to see if the file needs syncing
(which will be set to false if we lost the race to a athread that closed the file). The only other place safe to flush is on the
write path, but that would adversely impact write performance.

- cleaned up the logic in ESW.hasCapacityFor()

Further, I added the beginnings of test clases/infrastructure for testing the encryption code paths.