SASI: support for flushing partial SAs
Adds support for limiting the amount of memory used by a
SASI.PerSSTableIndexWriter.ColumnIndex. If the limit is
hit a "partial" SA is written to disk. The parts are then
stitched together when the writing of the sstable is complete
and all parts have been written. Parts may be written concurrently.

For memtable flushes the maximum memory is equal to
memtable_total_space_in_mb. For compacted sstables,
the default maximum is 15% of memtable_total_space_in_mb.
The default for compacted sstables can be overriden on a per-column
basis by setting the 'max_compaction_flush_memory_in_mb' index option
on the column.