Use Caffeine for key, counter, & Guava on-heap caches
Caffeine is a Java 8 cache by the author of ConcurrentLinkedHashMap,
which is what was previously used. That cache remains as a fallback
if needed, but is deprecated and scheduled for removal.

CLHM uses the LRU policy due to its simplicity, decent hit rate, and
known characteristics. Caffeine uses a near optimal policy, W-TinyLFU,
which has a significantly higher hit rate. In particular the hit rate
in the paper shows a substantial gain on database and search workloads.

The performance between the two caches should be similar, with some
possible gains in write throughput.

Caffeine is the successor to Guava's cache as well and those usages were
migrated. The performance should be significantly better, but the impact
small due to the usages not being heavily contended.

Caffeine: https://github.com/ben-manes/caffeine
W-TinyLFU: http://arxiv.org/pdf/1512.00727.pdf
Overview: https://www.voxxed.com/blog/2015/12/add-a-boost-of-caffeine-to-your-java
Design: http://highscalability.com/blog/2016/1/25/design-of-a-modern-cache.html